async/requirements.txt
python-socketio==4.5.1
aiohttp[speedups]==3.6.2
aiohttp-devtools==0.13.1
